fix: provide default avatar_url when creating a new user
This commit fixes a bug where user creation was failing with a 400 error.

The issue was caused by a null value being inserted into the vatar_url column of the people table, which is a non-nullable column.

To fix this, I've updated the ensureUserExists function to provide a default vatar_url when creating a new user.